+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.1] - 2021-05-15
+### Changed
+- Changed the required oo.library version to 1.12
+
+### Fixed
+- Fixed a bug on requesting radio stations without using the filters
+  https://www.amigans.net/modules/xforum/viewtopic.php?post_id=125112#forumpost125112
+  https://www.amigans.net/modules/xforum/viewtopic.php?post_id=125171#forumpost125171
+
 ## [1.1.0] - 2021-05-13
 ### Added
 - Added a message requester for messages to the user